refactor: split memory-storage::metadata.rs
